finish delete list components

+++ frontend/web/src/app/dialogs/componentslist/componentslist.component.ts
@@ -14,7 +14,7 @@ import {ErrorService} from '../../services/error.service';
   styleUrls: ['./componentslist.component.scss']
 })
 export class ComponentslistComponent implements AfterViewInit {
-  displayedColumns: string[] = ['nombre'];
+  displayedColumns: string[] = ['nombre', 'actions'];
   dataSource = new MatTableDataSource<BaseComponent>();
 
   constructor(
@@ -31,11 +31,13 @@ export class ComponentslistComponent implements AfterViewInit {
 
   getCustomList() {
     this.orderService.getCustomList('?productId=' + this.data.code).subscribe( result => {
+      console.log('response', result.response)
       this.dataSource.data = result.response;
     });
   }
 
   selectComponent(element: BaseComponent) {
+    console.log('element: ', element)
     this.dataService.presentToastCustom(Messages.confirmReplaceWithListComponents, 'question', '', true, true).then( (res: any) => {
       if (res.isConfirmed) {
         this.dialogRef.close({componentes: element.components});
@@ -44,9 +46,15 @@ export class ComponentslistComponent implements AfterViewInit {
   }
   removeCustomList(element: BaseComponent) {
     console.log('element: ', element)
-    this.orderService.deleteCustomList({productId: element.productId, name: element.name}).subscribe( deleteListResult => {
-      console.log('deleteResult: ', deleteListResult)
-    }, error =>  this.errorService.httpError(error));
+
+    this.dataService.presentToastCustom(`${Messages.removeListComponents} ${element.name.toUpperCase()}?`,
+        'question', '', true, true).then( (res: any) => {
+      if (res.isConfirmed) {
+        this.orderService.deleteCustomList({productId: element.productId, name: element.name}).subscribe( deleteListResult => {
+          console.log('deleteResult: ', deleteListResult)
+          this.getCustomList();
+        }, error =>  this.errorService.httpError(error));
+      }});
   }
 
 }
